backend/requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
qrcode[pil]>=7.4.2
Pillow>=10.0.0
pyzbar>=0.1.9
opencv-python-headless>=4.8.0.76
requests>=2.31.0
beautifulsoup4>=4.12.2
python-multipart>=0.0.6
